Fixed instantiation of SimpleFTPServer
The way that was used before did lead to 'handler' being a class variable.
When multiple servers were started the 'handler' and thus the 'authorizer' was overwritten for all instances, which did lead to an 'ftplib.error_perm: 530 Authentication failed' errors.
Also, set 'ignore_errors=True' when removing the temp downloaddir and used 'ensure(<path>, dir=True)' instead of 'mkdir(<path>)' in the tests which caused tests to fail with py39 on windows.